refactor: adding button factory
