+++ cinema-planner/cinema-planner/Views/HeroHeaderUIView.swift
@@ -9,28 +9,9 @@ import UIKit
 
 class HeroHeaderUIView: UIView {
     
-    private let downloadButton: UIButton = {
-        let button = UIButton()
-        button.setTitle("Download", for: .normal)
-        button.layer.borderColor = UIColor.white.cgColor
-        button.layer.borderWidth = 1
-        button.layer.cornerRadius = 5
-        button.translatesAutoresizingMaskIntoConstraints = false
-        
-        return button
-    }()
+    private let downloadButton: UIButton = makeButton(withText: "Download", withButtonType: .normal)
     
-    private let playButton: UIButton = {
-        
-        let button = UIButton()
-        button.setTitle("Play", for: .normal)
-        button.layer.borderColor = UIColor.white.cgColor
-        button.layer.borderWidth = 1
-        button.layer.cornerRadius = 5
-        button.translatesAutoresizingMaskIntoConstraints = false
-        
-        return button
-    }()
+    private let playButton: UIButton = makeButton(withText: "Play", withButtonType: .normal)
     
     private let heroImageView: UIImageView = {
